refactor cloudwatch filename
