+++ javascriptv3/example_code/cloudwatch-events/src/libs/cloudWatchEventsClient.js
