+++ javascriptv3/example_code/cloudwatch-events/src/putEvents.js
