+++ javascriptv3/example_code/cloudwatch-events/src/putTargets.js
