+++ javascriptv3/example_code/cloudwatch-logs/package.json
@@ -6,17 +6,15 @@
   "author": "Brian Murray <brmur@amazon.com>, Alex Forsyth <alex-git@amazon.com>",
   "license": "Apache 2.0",
   "dependencies": {
-    "@aws-sdk/client-cloudwatch": "^3.3.0",
+    "@aws-sdk/client-cloudwatch-logs": "^3.3.0",
     "@aws-sdk/node-http-handler": "^3.3.0",
     "@aws-sdk/types": "^3.3.0",
     "node-fetch": "^2.6",
     "jest": "^26.6.3"
   },
-  "scripts": {
-    "test": "jest"
-  },
   "devDependencies": {
     "@types/node": "^14.0.23",
     "typescript": "^4.0.2"
-  }
+  },
+  "type": "module"
 }
