+++ javascriptv3/example_code/cloudwatch-logs/src/describeSubscriptionFilters.js
