+++ javascriptv3/example_code/cloudwatch-logs/src/libs/cloudwatchLogsClient.js
