--- javascriptv3/example_code/cloudwatch/src/metadata.yaml
@@ -1,52 +1,0 @@
----
-# Amazon Cloudwatch basic usage examples
-files:
-  - path: .libs/cwClient.js
-    services:
-      - cloudwatch
-  - path: .libs/cweClient.js
-      services:
-      - cloudwatch
-  - path: .libs/cwlClient.js
-      services:
-      - cloudwatch
-  - path: cw_deletealarms.js
-    services:
-      - cloudwatch
-  - path: cw_describealarms.js
-    services:
-      - cloudwatch
-  - path: cw_disablealarmactions.js
-    services:
-      - cloudwatch
-  - path: cw_enablealarmactions.js
-    services:
-      - cloudwatch
-  - path: cw_listmetrics.js
-    services:
-      - cloudwatch
-  - path: cw_putmetricalarm.js
-    services:
-      - cloudwatch
-  - path: cw_putmetricdata.js
-    services:
-      - cloudwatch
-  - path: cwe_putevents.js
-    services:
-      - cloudwatch
-  - path: cwe_putrule.js
-    services:
-      - cloudwatch
-  - path: cwe_puttargets.js
-    services:
-      - cloudwatch
-  - path: cwl_deletesubscriptionfilter.js
-    services:
-      - cloudwatch
-  - path: cwl_describesubscriptionfilters.js
-    services:
-      - cloudwatch
-  - path: cwl_putsubscriptionfilter.js
-    services:
-      - cloudwatch
-...
