+++ javascriptv3/example_code/tests/cloudwatch-events/package.json
@@ -1,13 +1,12 @@
 {
-  "name": "aws-sdk-v3-cloudwatch-examples",
+  "name": "aws-sdk-v3-cloudwatch-events-examples",
   "version": "1.0.0",
   "main": "index.js",
   "repository": "git@github.com/awsdocs/aws-doc-sdk-examples/tree/master/javascriptv3/example_code/cloudwatch.git",
   "author": "Brian Murray <brmur@amazon.com>, Alex Forsyth <alex-git@amazon.com>",
   "license": "Apache 2.0",
   "dependencies": {
-    "@aws-sdk/client-cloudwatch": "^3.3.0",
-    "@aws-sdk/node-http-handler": "^3.3.0",
+    "@aws-sdk/client-cloudwatch-events": "^3.3.0",
     "@aws-sdk/types": "^3.3.0",
     "node-fetch": "^2.6",
     "jest": "^26.6.3"
