+++ javascriptv3/example_code/tests/cloudwatch-logs/cwl_deletesubscriptionfilter.test.js
@@ -1,8 +1,8 @@
 // Get service clients module and commands using CommonJS syntax.
-const { run, params } = require("../../cloudwatch/src/describeSubscriptionFilters");
-const { cwlClient } = require("../../cloudwatch/src/libs/cloudwatchLogsClient");
+const { run, params } = require("../../../example_code-logs/cloudwatch/src/deleteSubscriptionFilter.js");
+const { cwlClient } = require("../../cloudwatch-logs/src/libs/cloudwatchLogsClient.js");
 
-jest.mock("../../cloudwatch/src/libs/cloudwatchLogsClient.js");
+jest.mock("../../cloudwatch-logs/src/libs/cloudwatchLogsClient.js");
 
 describe("@aws-sdk/client-cloudwatch-logs mock", () => {
     it("should successfully mock CloudWatch Logs client", async () => {
