+++ javascriptv3/example_code/tests/cloudwatch-logs/cwl_describesubscriptionfilters.test.js
@@ -1,8 +1,8 @@
 // Get service clients module and commands using CommonJS syntax.
-const { run, params } = require("../../../example_code/cloudwatch/src/deleteSubscriptionFilter.js");
-const { cwlClient } = require("../../cloudwatch/src/libs/cloudwatchLogsClient.js");
+const { run, params } = require("../../cloudwatch-logs/src/describeSubscriptionFilters");
+const { cwlClient } = require("../../cloudwatch-logs/src/libs/cloudwatchLogsClient");
 
-jest.mock("../../cloudwatch/src/libs/cloudwatchLogsClient.js");
+jest.mock("../../cloudwatch-logs/src/libs/cloudwatchLogsClient.js");
 
 describe("@aws-sdk/client-cloudwatch-logs mock", () => {
     it("should successfully mock CloudWatch Logs client", async () => {
